Changed all instances of (unsigned) long to uint32_t, int32_t. Multiplayer was not working on AMD64 because the stats checksumming was realying on 32-bit longs.
git-svn-id: file:///home/tjaden/{repo}/redpixel/trunk@1089 ec8ac367-a008-0410-9901-8f3136df4969

+++ src/demo.c
@@ -93,7 +93,7 @@ void demo_read_header()
 }
 
 
-long demo_read_seed()
+int32_t demo_read_seed()
 {
     return pack_igetl(fp);    
 }
@@ -202,7 +202,7 @@ void demo_write_close()
 }
 
 
-void demo_write_set_rng_seed(long seed)
+void demo_write_set_rng_seed(int32_t seed)
 {
     if (!fp) return;
