+++ src/include/demo.h
@@ -13,7 +13,7 @@ extern char demo_description[];
 
 int demo_read_open(char *filename);
 void demo_read_header();
-long demo_read_seed();
+int32_t demo_read_seed();
 int demo_read_next_packet_type();
 void demo_read_packet(void *p, int len);
 void demo_read_string(char *p, int len);
@@ -24,7 +24,7 @@ int demo_write_open(char *filename, int players, char *names[]);
 int demo_is_recording();
 void demo_write_close();
 
-void demo_write_set_rng_seed(long seed);
+void demo_write_set_rng_seed(int32_t seed);
 void demo_write_change_map(char *filename);
 void demo_write_frame_data(void *packet, int len);
 void demo_write_player_inputs();
