+++ src/include/rnd.h
@@ -1,6 +1,6 @@
 int rnd();
-void srnd(unsigned long seed);
+void srnd(uint32_t seed);
 
 int irnd();
-void sirnd(unsigned long seed);
+void sirnd(uint32_t seed);
