+++ src/include/skhelp.h
@@ -1,3 +1,5 @@
-void send_long(long val);
-long recv_long();
+#include <stdint.h>
+
+void send_long(int32_t val);
+int32_t recv_long();
 int receive_string(char *dest);
